nim: fix infer_ops.py case
Fixes https://github.com/adsharma/py2many/issues/16

+++ tests/expected/infer_ops.nim
@@ -16,37 +16,37 @@ proc foo() =
   let _ = -(a)
 
 proc add1(x: int8, y: int8): int16 =
-  return (x + y)
+  return int16((x + y))
 
 proc add2(x: int16, y: int16): int32 =
-  return (x + y)
+  return int32((x + y))
 
 proc add3(x: int32, y: int32): int64 =
-  return (x + y)
+  return int64((x + y))
 
 proc add4(x: int64, y: int64): int64 =
   return (x + y)
 
 proc add5(x: uint8, y: uint8): uint16 =
-  return (x + y)
+  return uint16((x + y))
 
 proc add6(x: uint16, y: uint16): uint32 =
-  return (x + y)
+  return uint32((x + y))
 
 proc add7(x: uint32, y: uint32): uint64 =
-  return (x + y)
+  return uint64((x + y))
 
 proc add8(x: uint64, y: uint64): uint64 =
   return (x + y)
 
 proc add9(x: int8, y: uint16): uint32 =
-  return (x + int8(y))
+  return uint32((uint16(x) + y))
 
 proc sub(x: int8, y: int8): int8 =
   return (x - y)
 
 proc mul(x: int8, y: int8): int16 =
-  return (x * y)
+  return int16((x * y))
 
 proc fadd1(x: int8, y: float): float =
   return (float(x) + y)
